refactor: chart util
Chart 컴포넌트 안에 구현한 getChartDate를
util/chart 로 이동시킴
